\name{Squared Exponential Covariance}
\alias{C2EC}
\title{Squared Exponential Covariance}
\usage{
squared_exponential_covariance(x1, y1, x2, y2, range, sill)
}
\description{
 In this function, x1, y1, x2, and y2 are the coordinates of two points for which we want to calculate the covariance. range is the spatial range parameter, and sill is the covariance value at the range distance.
 	range is the spatial range parameter, and sill is the covariance value at the range distance.
 	Note that this is a basic example of a squared exponential covariance function, and there are many more complex covariance functions
 	available in external packages like gstat and spatial.
}
\examples{
# Example usage

}
