\name{Plot posterior constant covariance}
\alias{Plot CC}
\title{Plot posterior constant covariance}
\usage{
plot_posterior_const_covariance(sigma_sq, a, b, n, x)
}
\description{
This function takes four arguments:

   		 sigma_sq: The prior estimate of the covariance.

    		a: The prior shape parameter.

    		b: The prior rate parameter.

    		n: The number of observations.

    		x: The sum of squared residuals.
}
\examples{
# Define prior parameters
sigma_sq <- 1.5
a <- 2
b <- 1
# Define data parameters
n <- 10
x <- 20

# Call the function to plot the posterior
plot_posterior_const_covariance(sigma_sq, a, b, n, x)
}
