\name{Polynomial Covariance}
\alias{CPC}
\title{Polynomial Covariance}
\usage{
polynomial_covariance(x1, y1, x2, y2, range, sill, exponent)
}
\description{
In this function, x1, y1, x2, and y2 are the coordinates of two points for which we want to calculate the covariance. sill is the covariance value at the range distance, and exponent is the degree of the polynomial (i.e., the order of the polynomial plus one).
	Note that this is a basic example of a polynomial covariance function, and there are many more complex covariance functions available in
 	external packages like gstat and spatial.
}
\examples{
# Example usage

}
